+++ src/init-action.ts
@@ -40,7 +40,7 @@ import {
   initConfig,
   runInit,
 } from "./init";
-import { Language } from "./languages";
+import { KnownLanguage } from "./languages";
 import { getActionsLogger, Logger } from "./logging";
 import {
   downloadOverlayBaseDatabaseFromCache,
@@ -359,6 +359,39 @@ async function run() {
     }
     core.endGroup();
 
+    // Set CODEQL_ENABLE_EXPERIMENTAL_FEATURES for Rust. We need to set this environment
+    // variable before initializing the config, otherwise Rust analysis will not be
+    // enabled.
+    if (
+      // Only enable Rust analysis if the user has explicitly requested it - don't
+      // enable it via language autodetection.
+      configUtils
+        .getRawLanguagesNoAutodetect(getOptionalInput("languages"))
+        .includes(KnownLanguage.rust)
+    ) {
+      const feat = Feature.RustAnalysis;
+      const minVer = featureConfig[feat].minimumVersion as string;
+      const envVar = "CODEQL_ENABLE_EXPERIMENTAL_FEATURES";
+      // if in default setup, it means the feature flag was on when rust was enabled
+      // if the feature flag gets turned off, let's not have rust analysis throwing a configuration error
+      // in that case rust analysis will be disabled only when default setup is refreshed
+      if (isDefaultSetup() || (await features.getValue(feat, codeql))) {
+        core.exportVariable(envVar, "true");
+      }
+      if (process.env[envVar] !== "true") {
+        throw new ConfigurationError(
+          `Experimental and not officially supported Rust analysis requires setting ${envVar}=true in the environment`,
+        );
+      }
+      const actualVer = (await codeql.getVersion()).version;
+      if (semver.lt(actualVer, minVer)) {
+        throw new ConfigurationError(
+          `Experimental rust analysis is supported by CodeQL CLI version ${minVer} or higher, but found version ${actualVer}`,
+        );
+      }
+      logger.info("Experimental rust analysis enabled");
+    }
+
     config = await initConfig({
       languagesInput: getOptionalInput("languages"),
       queriesInput: getOptionalInput("queries"),
@@ -486,7 +519,7 @@ async function run() {
     }
 
     if (
-      config.languages.includes(Language.swift) &&
+      config.languages.includes(KnownLanguage.swift) &&
       process.platform === "linux"
     ) {
       logger.warning(
@@ -495,7 +528,7 @@ async function run() {
     }
 
     if (
-      config.languages.includes(Language.go) &&
+      config.languages.includes(KnownLanguage.go) &&
       process.platform === "linux"
     ) {
       try {
@@ -553,7 +586,7 @@ async function run() {
         if (e instanceof FileCmdNotFoundError) {
           addDiagnostic(
             config,
-            Language.go,
+            KnownLanguage.go,
             makeDiagnostic(
               "go/workflow/file-program-unavailable",
               "The `file` program is required on Linux, but does not appear to be installed",
@@ -603,7 +636,7 @@ async function run() {
       core.exportVariable(kotlinLimitVar, "2.1.20");
     }
 
-    if (config.languages.includes(Language.cpp)) {
+    if (config.languages.includes(KnownLanguage.cpp)) {
       const envVar = "CODEQL_EXTRACTOR_CPP_TRAP_CACHING";
       if (process.env[envVar]) {
         logger.info(
@@ -622,7 +655,7 @@ async function run() {
     }
 
     // Set CODEQL_EXTRACTOR_CPP_BUILD_MODE_NONE
-    if (config.languages.includes(Language.cpp)) {
+    if (config.languages.includes(KnownLanguage.cpp)) {
       const bmnVar = "CODEQL_EXTRACTOR_CPP_BUILD_MODE_NONE";
       const value =
         process.env[bmnVar] ||
@@ -633,7 +666,7 @@ async function run() {
 
     // For rust: set CODEQL_ENABLE_EXPERIMENTAL_FEATURES, unless codeql already supports rust without it
     if (
-      config.languages.includes(Language.rust) &&
+      config.languages.includes(KnownLanguage.rust) &&
       !(await codeql.resolveLanguages()).rust
     ) {
       const feat = Feature.RustAnalysis;